// serverless/interfaces/step-function.interface.ts
export interface StepFunction {
    stateMachines: StateMachines;
    validate?: boolean;
}

type StateMachines = {
    [stateMachine: string]: {
        id?: string;
        name: string;
        definition: Definition;
        role?: string;
        tracingConfig?: TracingConfig;
    };
};

type TracingConfig = {
    enabled: boolean;
};

type Definition = {
    Comment?: string;
    StartAt: string;
    States: States;
};